+++ src/app/shared/utils/toast.service.ts
@@ -1,0 +1,18 @@
+import { Injectable } from '@angular/core';
+import { ToastController } from '@ionic/angular';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class ToastService {
+  constructor(private toastCtrl: ToastController) {}
+
+  async presentToast(message: string) {
+    const toast = await this.toastCtrl.create({
+      message: message,
+      duration: 2000,
+      position: 'bottom',
+    });
+    toast.present();
+  }
+}
